Move functions to object class
