--- cache/src/main/kotlin/gg/rsmod/cache/struct/Index.kt
@@ -1,6 +1,0 @@
-package gg.rsmod.cache.struct
-
-/**
- * @author Tom <rspsmods@gmail.com>
- */
-data class Index(val id: Int)
